Generate more realistic chart versions
The distribution of version numbers for real-world software packages
follows Zipf's Law[citation needed], so use the Zipfian RNG conveniently
provided by the Go standard library to generate realistic version
numbers. Smaller numbers have a higher probability of being generated,
so generated versions are more likely to have clumps sharing the same
major or minor component, which is more realistic for testing
performance. Generate more realistic pre-release versions, with a
variable number of pre-release parts and a mix of numeric and
alphanumeric parts, which is more like real-world pre-release versions
and better excercises code paths which deal with processing and sorting
prerelease versions.

Generate short chart names and small numbers in version components,
which is easier for humans to read and compare visually when debugging.

Simplify the worker-pool code, taking advantage of the new
concurrency-safe random chart version generator. Simplify error handling
by handling HTTP 422 responses as expected and bailing on any other
error.

Add command-line flags and graceful cancelation on interrupt. Drop the
username and password configuration flags as those can be specified in
the URL itself. The Go HTTP client sets the basic-auth request header
itself when a username and password are specified in the URL.